X-1: Abstracted deployment tasks into seperate files and used gulp registry.
