+++ deployment/tasks/DisableJekyll.ts
@@ -1,0 +1,7 @@
+import {task} from "gulp";
+import {open} from "fs";
+
+task("DisableJekyll", done => {
+    const callback = (err: Error) => done(err ? err : undefined);
+    open("./docs/.nojekyll", "w", callback);
+});
