+++ docs/classes/limitdefault.html
@@ -132,7 +132,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides <a href="sequencedefault.html">SequenceDefault</a>.<a href="sequencedefault.html#constructor">constructor</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/defaults/LimitDefault.ts#L8">defaults/LimitDefault.ts:8</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/defaults/LimitDefault.ts#L8">defaults/LimitDefault.ts:8</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -154,7 +154,7 @@
 					<div class="tsd-signature tsd-kind-icon">amount<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/defaults/LimitDefault.ts#L8">defaults/LimitDefault.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/defaults/LimitDefault.ts#L8">defaults/LimitDefault.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -165,7 +165,7 @@
 					<aside class="tsd-sources">
 						<p>Inherited from <a href="sequencedefault.html">SequenceDefault</a>.<a href="sequencedefault.html#coalescingoperator">coalescingOperator</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceDefault.ts#L11">SequenceDefault.ts:11</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceDefault.ts#L11">SequenceDefault.ts:11</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -177,7 +177,7 @@
 						<p>Implementation of <a href="../interfaces/isequencedefault.html">ISequenceDefault</a>.<a href="../interfaces/isequencedefault.html#default">default</a></p>
 						<p>Inherited from <a href="sequencedefault.html">SequenceDefault</a>.<a href="sequencedefault.html#default">default</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceDefault.ts#L10">SequenceDefault.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceDefault.ts#L10">SequenceDefault.ts:10</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -189,7 +189,7 @@
 						<p>Implementation of <a href="../interfaces/isequencedefault.html">ISequenceDefault</a>.<a href="../interfaces/isequencedefault.html#sortings">sortings</a></p>
 						<p>Inherited from <a href="sequencedefault.html">SequenceDefault</a>.<a href="sequencedefault.html#sortings">sortings</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceDefault.ts#L12">SequenceDefault.ts:12</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceDefault.ts#L12">SequenceDefault.ts:12</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -208,7 +208,7 @@
 								<p>Implementation of <a href="../interfaces/isequencedefault.html">ISequenceDefault</a>.<a href="../interfaces/isequencedefault.html#stringify">stringify</a></p>
 								<p>Overrides <a href="sequencedefault.html">SequenceDefault</a>.<a href="sequencedefault.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/defaults/LimitDefault.ts#L16">defaults/LimitDefault.ts:16</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/defaults/LimitDefault.ts#L16">defaults/LimitDefault.ts:16</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -226,7 +226,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -244,7 +244,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -268,7 +268,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencedefault.html">SequenceDefault</a>.<a href="sequencedefault.html#stringifydefaulttype">stringifyDefaultType</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceDefault.ts#L29">SequenceDefault.ts:29</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceDefault.ts#L29">SequenceDefault.ts:29</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
