+++ docs/classes/logicalconditional.html
@@ -132,7 +132,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#constructor">constructor</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/LogicalConditional.ts#L10">conditionals/LogicalConditional.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/LogicalConditional.ts#L10">conditionals/LogicalConditional.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -162,7 +162,7 @@
 						<p>Implementation of <a href="../interfaces/isequenceconditional.html">ISequenceConditional</a>.<a href="../interfaces/isequenceconditional.html#column">column</a></p>
 						<p>Inherited from <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#column">column</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceConditional.ts#L9">SequenceConditional.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceConditional.ts#L9">SequenceConditional.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -174,7 +174,7 @@
 						<p>Implementation of <a href="../interfaces/isequenceconditional.html">ISequenceConditional</a>.<a href="../interfaces/isequenceconditional.html#conditional">conditional</a></p>
 						<p>Inherited from <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#conditional">conditional</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceConditional.ts#L8">SequenceConditional.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceConditional.ts#L8">SequenceConditional.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -184,7 +184,7 @@
 					<div class="tsd-signature tsd-kind-icon">logical<wbr>Operator<span class="tsd-signature-symbol">:</span> <a href="../enums/logicaloperator.html" class="tsd-signature-type">LogicalOperator</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/LogicalConditional.ts#L9">conditionals/LogicalConditional.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/LogicalConditional.ts#L9">conditionals/LogicalConditional.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -194,7 +194,7 @@
 					<div class="tsd-signature tsd-kind-icon">value<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/LogicalConditional.ts#L10">conditionals/LogicalConditional.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/LogicalConditional.ts#L10">conditionals/LogicalConditional.ts:10</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -213,7 +213,7 @@
 								<p>Implementation of <a href="../interfaces/isequenceconditional.html">ISequenceConditional</a>.<a href="../interfaces/isequenceconditional.html#stringify">stringify</a></p>
 								<p>Overrides <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/LogicalConditional.ts#L19">conditionals/LogicalConditional.ts:19</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/LogicalConditional.ts#L19">conditionals/LogicalConditional.ts:19</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -231,7 +231,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -249,7 +249,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -273,7 +273,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#stringifyconditional">stringifyConditional</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceConditional.ts#L23">SequenceConditional.ts:23</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceConditional.ts#L23">SequenceConditional.ts:23</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -296,7 +296,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/LogicalConditional.ts#L24">conditionals/LogicalConditional.ts:24</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/LogicalConditional.ts#L24">conditionals/LogicalConditional.ts:24</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
