+++ docs/classes/parenthesesconditional.html
@@ -130,7 +130,7 @@
 							<aside class="tsd-sources">
 								<p>Overrides <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#constructor">constructor</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/ParenthesesConditional.ts#L8">conditionals/ParenthesesConditional.ts:8</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/ParenthesesConditional.ts#L8">conditionals/ParenthesesConditional.ts:8</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -160,7 +160,7 @@
 						<p>Implementation of <a href="../interfaces/isequenceconditional.html">ISequenceConditional</a>.<a href="../interfaces/isequenceconditional.html#column">column</a></p>
 						<p>Inherited from <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#column">column</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceConditional.ts#L9">SequenceConditional.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceConditional.ts#L9">SequenceConditional.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -172,7 +172,7 @@
 						<p>Implementation of <a href="../interfaces/isequenceconditional.html">ISequenceConditional</a>.<a href="../interfaces/isequenceconditional.html#conditional">conditional</a></p>
 						<p>Inherited from <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#conditional">conditional</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceConditional.ts#L8">SequenceConditional.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceConditional.ts#L8">SequenceConditional.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -182,7 +182,7 @@
 					<div class="tsd-signature tsd-kind-icon">values<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">any</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/ParenthesesConditional.ts#L8">conditionals/ParenthesesConditional.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/ParenthesesConditional.ts#L8">conditionals/ParenthesesConditional.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -201,7 +201,7 @@
 								<p>Implementation of <a href="../interfaces/isequenceconditional.html">ISequenceConditional</a>.<a href="../interfaces/isequenceconditional.html#stringify">stringify</a></p>
 								<p>Overrides <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/conditionals/ParenthesesConditional.ts#L16">conditionals/ParenthesesConditional.ts:16</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/conditionals/ParenthesesConditional.ts#L16">conditionals/ParenthesesConditional.ts:16</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -219,7 +219,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -237,7 +237,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -261,7 +261,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequenceconditional.html">SequenceConditional</a>.<a href="sequenceconditional.html#stringifyconditional">stringifyConditional</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceConditional.ts#L23">SequenceConditional.ts:23</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceConditional.ts#L23">SequenceConditional.ts:23</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
