+++ docs/classes/selectionoperation.html
@@ -129,7 +129,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/operations/SelectionOperation.ts#L10">operations/SelectionOperation.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/operations/SelectionOperation.ts#L10">operations/SelectionOperation.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -153,7 +153,7 @@
 						<p>Implementation of <a href="../interfaces/isequenceoperation.html">ISequenceOperation</a>.<a href="../interfaces/isequenceoperation.html#columns">columns</a></p>
 						<p>Overrides <a href="sequenceoperation.html">SequenceOperation</a>.<a href="sequenceoperation.html#columns">columns</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/operations/SelectionOperation.ts#L10">operations/SelectionOperation.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/operations/SelectionOperation.ts#L10">operations/SelectionOperation.ts:10</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -165,7 +165,7 @@
 						<p>Implementation of <a href="../interfaces/isequenceoperation.html">ISequenceOperation</a>.<a href="../interfaces/isequenceoperation.html#operation">operation</a></p>
 						<p>Overrides <a href="sequenceoperation.html">SequenceOperation</a>.<a href="sequenceoperation.html#operation">operation</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/operations/SelectionOperation.ts#L9">operations/SelectionOperation.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/operations/SelectionOperation.ts#L9">operations/SelectionOperation.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -184,7 +184,7 @@
 								<p>Implementation of <a href="../interfaces/isequenceoperation.html">ISequenceOperation</a>.<a href="../interfaces/isequenceoperation.html#stringify">stringify</a></p>
 								<p>Overrides <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/operations/SelectionOperation.ts#L18">operations/SelectionOperation.ts:18</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/operations/SelectionOperation.ts#L18">operations/SelectionOperation.ts:18</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -202,7 +202,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -220,7 +220,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -244,7 +244,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequenceoperation.html">SequenceOperation</a>.<a href="sequenceoperation.html#stringifyoperation">stringifyOperation</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceOperation.ts#L11">SequenceOperation.ts:11</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceOperation.ts#L11">SequenceOperation.ts:11</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
