+++ docs/classes/sequencebuilder.html
@@ -124,7 +124,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#condition">condition</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L17">SequenceBuilder.ts:17</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L17">SequenceBuilder.ts:17</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -135,7 +135,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#grouping">grouping</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L22">SequenceBuilder.ts:22</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L22">SequenceBuilder.ts:22</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -146,7 +146,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#joins">joins</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L20">SequenceBuilder.ts:20</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L20">SequenceBuilder.ts:20</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -157,7 +157,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#limitation">limitation</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L23">SequenceBuilder.ts:23</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L23">SequenceBuilder.ts:23</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -168,7 +168,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#location">location</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L18">SequenceBuilder.ts:18</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L18">SequenceBuilder.ts:18</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -179,7 +179,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#operation">operation</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L19">SequenceBuilder.ts:19</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L19">SequenceBuilder.ts:19</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -190,7 +190,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#ordering">ordering</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L21">SequenceBuilder.ts:21</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L21">SequenceBuilder.ts:21</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -208,7 +208,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#groupby">groupBy</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L31">SequenceBuilder.ts:31</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L31">SequenceBuilder.ts:31</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -235,7 +235,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#limit">limit</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L37">SequenceBuilder.ts:37</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L37">SequenceBuilder.ts:37</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -259,7 +259,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#orderby">orderBy</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L25">SequenceBuilder.ts:25</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L25">SequenceBuilder.ts:25</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -286,7 +286,7 @@
 							<aside class="tsd-sources">
 								<p>Implementation of <a href="../interfaces/isequencebuilder.html">ISequenceBuilder</a>.<a href="../interfaces/isequencebuilder.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L47">SequenceBuilder.ts:47</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L47">SequenceBuilder.ts:47</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -303,7 +303,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L42">SequenceBuilder.ts:42</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L42">SequenceBuilder.ts:42</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -320,7 +320,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceBuilder.ts#L54">SequenceBuilder.ts:54</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceBuilder.ts#L54">SequenceBuilder.ts:54</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
