+++ docs/classes/sequencecolumn.html
@@ -127,7 +127,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceColumn.ts#L8">SequenceColumn.ts:8</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceColumn.ts#L8">SequenceColumn.ts:8</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -153,7 +153,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencecolumn.html">ISequenceColumn</a>.<a href="../interfaces/isequencecolumn.html#name">name</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceColumn.ts#L8">SequenceColumn.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceColumn.ts#L8">SequenceColumn.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -164,7 +164,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencecolumn.html">ISequenceColumn</a>.<a href="../interfaces/isequencecolumn.html#predicate">predicate</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceColumn.ts#L7">SequenceColumn.ts:7</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceColumn.ts#L7">SequenceColumn.ts:7</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -183,7 +183,7 @@
 								<p>Implementation of <a href="../interfaces/isequencecolumn.html">ISequenceColumn</a>.<a href="../interfaces/isequencecolumn.html#stringify">stringify</a></p>
 								<p>Overrides <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceColumn.ts#L17">SequenceColumn.ts:17</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceColumn.ts#L17">SequenceColumn.ts:17</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -201,7 +201,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -219,7 +219,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -242,7 +242,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceColumn.ts#L27">SequenceColumn.ts:27</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceColumn.ts#L27">SequenceColumn.ts:27</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
