+++ docs/classes/sequencecondition.html
@@ -128,7 +128,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceCondition.ts#L11">SequenceCondition.ts:11</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceCondition.ts#L11">SequenceCondition.ts:11</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -157,7 +157,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencecondition.html">ISequenceCondition</a>.<a href="../interfaces/isequencecondition.html#coalescingoperator">coalescingOperator</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceCondition.ts#L10">SequenceCondition.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceCondition.ts#L10">SequenceCondition.ts:10</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -168,7 +168,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencecondition.html">ISequenceCondition</a>.<a href="../interfaces/isequencecondition.html#condition">condition</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceCondition.ts#L9">SequenceCondition.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceCondition.ts#L9">SequenceCondition.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -179,7 +179,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequencecondition.html">ISequenceCondition</a>.<a href="../interfaces/isequencecondition.html#conditionals">conditionals</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceCondition.ts#L11">SequenceCondition.ts:11</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceCondition.ts#L11">SequenceCondition.ts:11</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -198,7 +198,7 @@
 								<p>Implementation of <a href="../interfaces/isequencecondition.html">ISequenceCondition</a>.<a href="../interfaces/isequencecondition.html#stringify">stringify</a></p>
 								<p>Overrides <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceCondition.ts#L21">SequenceCondition.ts:21</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceCondition.ts#L21">SequenceCondition.ts:21</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -216,7 +216,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -234,7 +234,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -257,7 +257,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceCondition.ts#L28">SequenceCondition.ts:28</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceCondition.ts#L28">SequenceCondition.ts:28</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
