+++ docs/classes/sequenceoperation.html
@@ -123,7 +123,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequenceoperation.html">ISequenceOperation</a>.<a href="../interfaces/isequenceoperation.html#columns">columns</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceOperation.ts#L9">SequenceOperation.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceOperation.ts#L9">SequenceOperation.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -134,7 +134,7 @@
 					<aside class="tsd-sources">
 						<p>Implementation of <a href="../interfaces/isequenceoperation.html">ISequenceOperation</a>.<a href="../interfaces/isequenceoperation.html#operation">operation</a></p>
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceOperation.ts#L8">SequenceOperation.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceOperation.ts#L8">SequenceOperation.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -153,7 +153,7 @@
 								<p>Implementation of <a href="../interfaces/isequenceoperation.html">ISequenceOperation</a>.<a href="../interfaces/isequenceoperation.html#stringify">stringify</a></p>
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L6">SequencePart.ts:6</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L6">SequencePart.ts:6</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -171,7 +171,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#tostring">toString</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L10">SequencePart.ts:10</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
@@ -189,7 +189,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="sequencepart.html">SequencePart</a>.<a href="sequencepart.html#stringifycoalescingoperator">stringifyCoalescingOperator</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequencePart.ts#L14">SequencePart.ts:14</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -212,7 +212,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/SequenceOperation.ts#L11">SequenceOperation.ts:11</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/SequenceOperation.ts#L11">SequenceOperation.ts:11</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
