+++ docs/enums/arrangement.html
@@ -88,7 +88,7 @@
 					<div class="tsd-signature tsd-kind-icon">Ascending<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Arrangement.ts#L2">enums/Arrangement.ts:2</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Arrangement.ts#L2">enums/Arrangement.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -98,7 +98,7 @@
 					<div class="tsd-signature tsd-kind-icon">Descending<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Arrangement.ts#L3">enums/Arrangement.ts:3</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Arrangement.ts#L3">enums/Arrangement.ts:3</a></li>
 						</ul>
 					</aside>
 				</section>
