+++ docs/enums/coalescingoperator.html
@@ -88,7 +88,7 @@
 					<div class="tsd-signature tsd-kind-icon">And<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/CoalescingOperator.ts#L3">enums/CoalescingOperator.ts:3</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/CoalescingOperator.ts#L3">enums/CoalescingOperator.ts:3</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -98,7 +98,7 @@
 					<div class="tsd-signature tsd-kind-icon">None<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/CoalescingOperator.ts#L2">enums/CoalescingOperator.ts:2</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/CoalescingOperator.ts#L2">enums/CoalescingOperator.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
