+++ docs/enums/location.html
@@ -87,7 +87,7 @@
 					<div class="tsd-signature tsd-kind-icon">From<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Location.ts#L2">enums/Location.ts:2</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Location.ts#L2">enums/Location.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
