+++ docs/enums/operation.html
@@ -87,7 +87,7 @@
 					<div class="tsd-signature tsd-kind-icon">Select<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Operation.ts#L2">enums/Operation.ts:2</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Operation.ts#L2">enums/Operation.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
