+++ docs/enums/predicate.html
@@ -89,7 +89,7 @@
 					<div class="tsd-signature tsd-kind-icon">Count<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Predicate.ts#L4">enums/Predicate.ts:4</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Predicate.ts#L4">enums/Predicate.ts:4</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -99,7 +99,7 @@
 					<div class="tsd-signature tsd-kind-icon">Distinct<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Predicate.ts#L3">enums/Predicate.ts:3</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Predicate.ts#L3">enums/Predicate.ts:3</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -109,7 +109,7 @@
 					<div class="tsd-signature tsd-kind-icon">None<span class="tsd-signature-symbol">:</span> </div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/enums/Predicate.ts#L2">enums/Predicate.ts:2</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/enums/Predicate.ts#L2">enums/Predicate.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
