+++ docs/interfaces/isequencebuilder.html
@@ -118,7 +118,7 @@
 					<div class="tsd-signature tsd-kind-icon">condition<span class="tsd-signature-symbol">:</span> <a href="isequencecondition.html" class="tsd-signature-type">ISequenceCondition</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L16">interfaces/ISequenceBuilder.ts:16</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L16">interfaces/ISequenceBuilder.ts:16</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -128,7 +128,7 @@
 					<div class="tsd-signature tsd-kind-icon">grouping<span class="tsd-signature-symbol">:</span> <a href="isequencedefault.html" class="tsd-signature-type">ISequenceDefault</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L22">interfaces/ISequenceBuilder.ts:22</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L22">interfaces/ISequenceBuilder.ts:22</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -138,7 +138,7 @@
 					<div class="tsd-signature tsd-kind-icon">joins<span class="tsd-signature-symbol">:</span> <a href="isequencejoin.html" class="tsd-signature-type">ISequenceJoin</a><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L18">interfaces/ISequenceBuilder.ts:18</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L18">interfaces/ISequenceBuilder.ts:18</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -148,7 +148,7 @@
 					<div class="tsd-signature tsd-kind-icon">limitation<span class="tsd-signature-symbol">:</span> <a href="../classes/limitdefault.html" class="tsd-signature-type">LimitDefault</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L24">interfaces/ISequenceBuilder.ts:24</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L24">interfaces/ISequenceBuilder.ts:24</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -158,7 +158,7 @@
 					<div class="tsd-signature tsd-kind-icon">location<span class="tsd-signature-symbol">:</span> <a href="isequencelocation.html" class="tsd-signature-type">ISequenceLocation</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L14">interfaces/ISequenceBuilder.ts:14</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L14">interfaces/ISequenceBuilder.ts:14</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -168,7 +168,7 @@
 					<div class="tsd-signature tsd-kind-icon">operation<span class="tsd-signature-symbol">:</span> <a href="isequenceoperation.html" class="tsd-signature-type">ISequenceOperation</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L12">interfaces/ISequenceBuilder.ts:12</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L12">interfaces/ISequenceBuilder.ts:12</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -178,7 +178,7 @@
 					<div class="tsd-signature tsd-kind-icon">ordering<span class="tsd-signature-symbol">:</span> <a href="isequencedefault.html" class="tsd-signature-type">ISequenceDefault</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L20">interfaces/ISequenceBuilder.ts:20</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L20">interfaces/ISequenceBuilder.ts:20</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -195,7 +195,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L28">interfaces/ISequenceBuilder.ts:28</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L28">interfaces/ISequenceBuilder.ts:28</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -221,7 +221,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L30">interfaces/ISequenceBuilder.ts:30</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L30">interfaces/ISequenceBuilder.ts:30</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -244,7 +244,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L26">interfaces/ISequenceBuilder.ts:26</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L26">interfaces/ISequenceBuilder.ts:26</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -270,7 +270,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceBuilder.ts#L32">interfaces/ISequenceBuilder.ts:32</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceBuilder.ts#L32">interfaces/ISequenceBuilder.ts:32</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
