+++ docs/interfaces/isequencecolumn.html
@@ -113,7 +113,7 @@
 					<div class="tsd-signature tsd-kind-icon">name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceColumn.ts#L8">interfaces/ISequenceColumn.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceColumn.ts#L8">interfaces/ISequenceColumn.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -123,7 +123,7 @@
 					<div class="tsd-signature tsd-kind-icon">predicate<span class="tsd-signature-symbol">:</span> <a href="../enums/predicate.html" class="tsd-signature-type">Predicate</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceColumn.ts#L6">interfaces/ISequenceColumn.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceColumn.ts#L6">interfaces/ISequenceColumn.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -141,7 +141,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="isequencepart.html">ISequencePart</a>.<a href="isequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
