+++ docs/interfaces/isequencecondition.html
@@ -114,7 +114,7 @@
 					<div class="tsd-signature tsd-kind-icon">coalescing<wbr>Operator<span class="tsd-signature-symbol">:</span> <a href="../enums/coalescingoperator.html" class="tsd-signature-type">CoalescingOperator</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceCondition.ts#L9">interfaces/ISequenceCondition.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceCondition.ts#L9">interfaces/ISequenceCondition.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -124,7 +124,7 @@
 					<div class="tsd-signature tsd-kind-icon">condition<span class="tsd-signature-symbol">:</span> <a href="../enums/condition.html" class="tsd-signature-type">Condition</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceCondition.ts#L8">interfaces/ISequenceCondition.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceCondition.ts#L8">interfaces/ISequenceCondition.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -134,7 +134,7 @@
 					<div class="tsd-signature tsd-kind-icon">conditionals<span class="tsd-signature-symbol">:</span> <a href="isequenceconditional.html" class="tsd-signature-type">ISequenceConditional</a><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceCondition.ts#L10">interfaces/ISequenceCondition.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceCondition.ts#L10">interfaces/ISequenceCondition.ts:10</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -152,7 +152,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="isequencepart.html">ISequencePart</a>.<a href="isequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
