+++ docs/interfaces/isequenceconditional.html
@@ -117,7 +117,7 @@
 					<div class="tsd-signature tsd-kind-icon">column<span class="tsd-signature-symbol">:</span> <a href="isequencecolumn.html" class="tsd-signature-type">ISequenceColumn</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceConditional.ts#L9">interfaces/ISequenceConditional.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceConditional.ts#L9">interfaces/ISequenceConditional.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -127,7 +127,7 @@
 					<div class="tsd-signature tsd-kind-icon">conditional<span class="tsd-signature-symbol">:</span> <a href="../enums/conditional.html" class="tsd-signature-type">Conditional</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceConditional.ts#L7">interfaces/ISequenceConditional.ts:7</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceConditional.ts#L7">interfaces/ISequenceConditional.ts:7</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -145,7 +145,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="isequencepart.html">ISequencePart</a>.<a href="isequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
