+++ docs/interfaces/isequencedefault.html
@@ -115,7 +115,7 @@
 					<div class="tsd-signature tsd-kind-icon">default<span class="tsd-signature-symbol">:</span> <a href="../enums/default.html" class="tsd-signature-type">Default</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceDefault.ts#L7">interfaces/ISequenceDefault.ts:7</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceDefault.ts#L7">interfaces/ISequenceDefault.ts:7</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -125,7 +125,7 @@
 					<div class="tsd-signature tsd-kind-icon">sortings<span class="tsd-signature-symbol">:</span> <a href="isequencesorting.html" class="tsd-signature-type">ISequenceSorting</a><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceDefault.ts#L8">interfaces/ISequenceDefault.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceDefault.ts#L8">interfaces/ISequenceDefault.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -143,7 +143,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="isequencepart.html">ISequencePart</a>.<a href="isequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
