+++ docs/interfaces/isequencelocation.html
@@ -112,7 +112,7 @@
 					<div class="tsd-signature tsd-kind-icon">location<span class="tsd-signature-symbol">:</span> <a href="../enums/location.html" class="tsd-signature-type">Location</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceLocation.ts#L6">interfaces/ISequenceLocation.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceLocation.ts#L6">interfaces/ISequenceLocation.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -130,7 +130,7 @@
 							<aside class="tsd-sources">
 								<p>Inherited from <a href="isequencepart.html">ISequencePart</a>.<a href="isequencepart.html#stringify">stringify</a></p>
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
