+++ docs/interfaces/isequencepart.html
@@ -144,7 +144,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
+									<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequencePart.ts#L3">interfaces/ISequencePart.ts:3</a></li>
 								</ul>
 							</aside>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4>
