+++ docs/interfaces/isequencesorting.html
@@ -102,7 +102,7 @@
 					<div class="tsd-signature tsd-kind-icon">arrangement<span class="tsd-signature-symbol">:</span> <a href="../enums/arrangement.html" class="tsd-signature-type">Arrangement</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceSorting.ts#L8">interfaces/ISequenceSorting.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceSorting.ts#L8">interfaces/ISequenceSorting.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -112,7 +112,7 @@
 					<div class="tsd-signature tsd-kind-icon">column<span class="tsd-signature-symbol">:</span> <a href="isequencecolumn.html" class="tsd-signature-type">ISequenceColumn</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/a14cf4c/src/interfaces/ISequenceSorting.ts#L6">interfaces/ISequenceSorting.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/Gallink/Sequence/blob/7eb8a85/src/interfaces/ISequenceSorting.ts#L6">interfaces/ISequenceSorting.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
