+++ package.json
@@ -28,6 +28,7 @@
     "chai": "^4.2.0",
     "gulp": "^4.0.0",
     "gulp-clean": "^0.4.0",
+    "gulp-hub": "^4.2.0",
     "gulp-typedoc": "^2.2.1",
     "gulp-typescript": "^5.0.0-alpha.3",
     "merge2": "^1.2.3",
